search: migrate structural search logic to own file
