+++ cmd/frontend/graphqlbackend/zoekt.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"math"
 	"net/url"
-	"regexp"
 	"regexp/syntax"
 	"strings"
 	"time"
@@ -16,7 +15,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/gituri"
-	"github.com/sourcegraph/sourcegraph/internal/lazyregexp"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	searchbackend "github.com/sourcegraph/sourcegraph/internal/search/backend"
 	"github.com/sourcegraph/sourcegraph/internal/symbols/protocol"
@@ -377,63 +375,6 @@ func parseRe(pattern string, filenameOnly bool, queryIsCaseSensitive bool) (zoek
 
 func fileRe(pattern string, queryIsCaseSensitive bool) (zoektquery.Q, error) {
 	return parseRe(pattern, true, queryIsCaseSensitive)
-}
-
-func splitOnHolesPattern() string {
-	word := `\w+`
-	whitespaceAndOptionalWord := `[ ]+(` + word + `)?`
-	holeAnything := `:\[` + word + `\]`
-	holeAlphanum := `:\[\[` + word + `\]\]`
-	holeWithPunctuation := `:\[` + word + `\.\]`
-	holeWithNewline := `:\[` + word + `\\n\]`
-	holeWhitespace := `:\[` + whitespaceAndOptionalWord + `\]`
-	return strings.Join([]string{
-		holeAnything,
-		holeAlphanum,
-		holeWithPunctuation,
-		holeWithNewline,
-		holeWhitespace,
-	}, "|")
-}
-
-var matchHoleRegexp = lazyregexp.New(splitOnHolesPattern())
-
-// Converts comby a structural pattern to a Zoekt regular expression query. It
-// converts whitespace in the pattern so that content across newlines can be
-// matched in the index. As an incomplete approximation, we use the regex
-// pattern .*? to scan ahead.
-// Example:
-// "ParseInt(:[args]) if err != nil" -> "ParseInt(.*)\s+if\s+err!=\s+nil"
-func StructuralPatToRegexpQuery(pattern string) (zoektquery.Q, error) {
-	substrings := matchHoleRegexp.Split(pattern, -1)
-	var children []zoektquery.Q
-	var pieces []string
-	for _, s := range substrings {
-		piece := regexp.QuoteMeta(s)
-		onMatchWhitespace := lazyregexp.New(`[\s]+`)
-		piece = onMatchWhitespace.ReplaceAllLiteralString(piece, `[\s]+`)
-		pieces = append(pieces, piece)
-	}
-
-	if len(pieces) == 0 {
-		return &zoektquery.Const{Value: true}, nil
-	}
-	rs := "(" + strings.Join(pieces, ")(.|\\s)*?(") + ")"
-	re, _ := syntax.Parse(rs, syntax.ClassNL|syntax.PerlX|syntax.UnicodeGroups)
-	children = append(children, &zoektquery.Regexp{
-		Regexp:        re,
-		CaseSensitive: true,
-		Content:       true,
-	})
-	return &zoektquery.And{Children: children}, nil
-}
-
-func StructuralPatToQuery(pattern string) (zoektquery.Q, error) {
-	regexpQuery, err := StructuralPatToRegexpQuery(pattern)
-	if err != nil {
-		return nil, err
-	}
-	return &zoektquery.Or{Children: []zoektquery.Q{regexpQuery}}, nil
 }
 
 func queryToZoektQuery(query *search.TextPatternInfo, isSymbol bool) (zoektquery.Q, error) {
